MOve prepareInputs to transaction agent level as preparation for refactor
