+++ examples/bri-3/src/bri/zeroKnowledgeProof/services/circuit/circuitService.interface.ts
@@ -3,13 +3,13 @@ import { Transaction } from '../../../transactions/models/transaction';
 
 export interface ICircuitService {
   createWitness(
-    tx: Transaction,
-    circuitName: string,
+    circuitInputs: object,
     pathToCircuit: string,
     pathToProvingKey: string,
     pathToVerificationKey: string,
     pathToWitnessCalculator?: string,
     pathToWitnessFile?: string,
   ): Promise<Witness>;
+
   verifyProofUsingWitness(witness: Witness): Promise<boolean>;
 }
